added roles and exceptions

+++ src/domain/services/user.service.ts
@@ -2,6 +2,8 @@ import { Injectable } from '@nestjs/common';
 import { User } from '../models/user.model';
 import { UserRepository } from '../repositories/user.repository';
 import { CreateUser } from '../models/create.user.model';
+import { DomainException } from '../exception/domain.exception';
+import { UserNotFoundExeption } from '../exception/user.not.found.exeption';
 
 @Injectable()
 export class UserService {
@@ -11,15 +13,21 @@ export class UserService {
     return this.userRepository.findAll();
   }
 
-  findById(id: string): Promise<User> {
-    return this.userRepository.findById(id);
+  async findById(id: string): Promise<User> {
+    const user: User = await this.userRepository.findById(id);
+    if(user === undefined) throw new UserNotFoundExeption('user with id: ' + id + ' not found');
+    return user;
   }
 
-  findByEmail(email: string): Promise<User> {
-    return this.userRepository.findByEmail(email);
+  async findByEmail(email: string): Promise<User> {
+    const user: User = await this.userRepository.findByEmail(email);
+    if(user === undefined) throw new UserNotFoundExeption('user with email: ' + email + ' not found');
+    return user;
   }
 
-  remove(id: string): Promise<void> {
+  async remove(id: string): Promise<void> {
+    const user: User = await this.userRepository.findById(id);
+    if(user === undefined) throw new UserNotFoundExeption('user with id: ' + id + ' not found');
     return this.userRepository.removeById(id);
   }
